ml_pipeline/requirements.txt
fastapi
uvicorn
scikit-learn
mlflow
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
numpy 